LayoutTests/imported/w3c/web-platform-tests/web-locks/frames.tentative.https-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.

Harness Error (TIMEOUT), message = null

TIMEOUT Window and Frame - shared mode Test timed out
NOTRUN Window and Frame - exclusive mode
NOTRUN Frame and Frame - exclusive mode
NOTRUN Terminated Frame with held lock
NOTRUN Navigated Frame with held lock
NOTRUN Navigated Frame with pending request
NOTRUN Removed Frame with pending request
NOTRUN Removed Frame as lock is granted
